requirements.txt
Flask==2.1.1
Werkzeug==2.0.2
facebook-scraper==0.2.59
pandas==2.1.4
gunicorn==21.2.0
pillow==10.2.0